requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
flask>=3.0.0
matplotlib>=3.9.0
lxml>=5.0.0
mpi4py>=3.1.5
scikit-learn>=1.3.0
networkx>=3.1.0
pyvis>=0.3.2
reportlab>=4.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pytest>=7.0.0
numpy>=1.24.1
scipy>=1.10.0
joblib>=1.3.0
pillow>=9.0.0